2051: rename class file, replace deprecated method, close writer by using "use", refactor code
