+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/bayern/EhrenamtskarteBayern.kt
@@ -4,9 +4,9 @@ import app.ehrenamtskarte.backend.stores.importer.ImportConfig
 import app.ehrenamtskarte.backend.stores.importer.addStep
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.DownloadLbe
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.FilterLbe
+import app.ehrenamtskarte.backend.stores.importer.bayern.steps.FilteredStoresCsvOutput
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.MapFromLbe
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.PostSanitizeFilter
-import app.ehrenamtskarte.backend.stores.importer.bayern.steps.WriteCsv
 import app.ehrenamtskarte.backend.stores.importer.bayern.types.FilteredStore
 import app.ehrenamtskarte.backend.stores.importer.common.steps.FilterDuplicates
 import app.ehrenamtskarte.backend.stores.importer.common.steps.SanitizeAddress
@@ -49,7 +49,7 @@ object EhrenamtskarteBayern : Pipeline {
             .addStep(FilterDuplicates(config, logger, filteredStores), logger) {
                 logger.info("== Filter duplicated data ==")
             }
-            .addStep(WriteCsv(config, filteredStores), logger) {
+            .addStep(FilteredStoresCsvOutput(config, filteredStores), logger) {
                 logger.info("== Write CSV Filtered Stores")
             }
             .addStep(Store(config, logger), logger) {
